refactor(ui): extract settings/data/{format,notice,parse}.ts — pure formatters + JSON parsers
